reblend ModeManager to VimState
- In vim-mode mode handling is done in VimState
- Then I extracted mode handling to ModeManager
- But now ModeManager’s task now get very small, I’m OK to re-blend it
again!
