+++ lib/base.js
@@ -392,11 +392,6 @@ class Base {
   emitDidFinishMutation(...args) { return this.vimState.emitDidFinishMutation(...args) } // prettier-ignore
   onDidFinishOperation(...args) { return this.vimState.onDidFinishOperation(...args) } // prettier-ignore
   onDidResetOperationStack(...args) { return this.vimState.onDidResetOperationStack(...args) } // prettier-ignore
-  onWillActivateMode(...args) { return this.vimState.onWillActivateMode(...args) } // prettier-ignore
-  onDidActivateMode(...args) { return this.vimState.onDidActivateMode(...args) } // prettier-ignore
-  preemptWillDeactivateMode(...args) { return this.vimState.preemptWillDeactivateMode(...args) } // prettier-ignore
-  onWillDeactivateMode(...args) { return this.vimState.onWillDeactivateMode(...args) } // prettier-ignore
-  onDidDeactivateMode(...args) { return this.vimState.onDidDeactivateMode(...args) } // prettier-ignore
   onDidCancelSelectList(...args) { return this.vimState.onDidCancelSelectList(...args) } // prettier-ignore
   subscribe(...args) { return this.vimState.subscribe(...args) } // prettier-ignore
   isMode(...args) { return this.vimState.isMode(...args) } // prettier-ignore
