+++ lib/occurrence-manager.js
@@ -199,7 +199,7 @@ module.exports = class OccurrenceManager {
 
       // To avoid selected occurrence ruined by normalization when deactivating blockwise
       if (this.vimState.isMode("visual", "blockwise")) {
-        this.vimState.modeManager.activate("visual", "characterwise")
+        this.vimState.activate("visual", "characterwise")
       }
 
       editor.setSelectedBufferRanges(rangesToSelect, {reversed})
