+++ lib/operation-stack.js
@@ -12,21 +12,15 @@ const Base = require("./base")
 //    if executable, then pop stack then execute(poppedOperation)
 //    if not executable, enter "operator-pending-mode"
 module.exports = class OperationStack {
-  get mode() {
-    return this.modeManager.mode
-  }
-  get submode() {
-    return this.modeManager.submode
-  }
+  get mode() { return this.vimState.mode } // prettier-ignore
+  get submode() { return this.vimState.submode } // prettier-ignore
 
   constructor(vimState) {
     this.vimState = vimState
     this.editor = vimState.editor
     this.editorElement = vimState.editorElement
-    this.modeManager = vimState.modeManager
-
+    
     this.vimState.onDidDestroy(() => this.destroy())
-
     this.reset()
   }
 
@@ -178,7 +172,7 @@ module.exports = class OperationStack {
 
     if (!top.isReady()) {
       if (this.mode === "normal" && top.isOperator()) {
-        this.modeManager.activate("operator-pending")
+        this.vimState.activate("operator-pending")
       }
       // Temporary set while command is running to achieve operation-specific keymap scopes
       this.addToClassList(top.getCommandNameWithoutPrefix() + "-pending")
@@ -203,7 +197,7 @@ module.exports = class OperationStack {
   cancel(operation) {
     if (this.mode === "operator-pending") {
       this.vimState.mutationManager.restoreCursorsToInitialPosition()
-      this.modeManager.activate("normal")
+      this.vimState.activate("normal")
     }
     this.finish(operation, true)
   }
@@ -226,7 +220,7 @@ module.exports = class OperationStack {
       const eolCursors = this.editor.getCursors().filter(cursor => cursor.isAtEndOfLine())
       eolCursors.forEach(cursor => this.vimState.utils.moveCursorLeft(cursor, {keepGoalColumn: true}))
     } else if (this.mode === "visual") {
-      this.modeManager.updateNarrowedState()
+      this.vimState.updateNarrowedState()
       this.vimState.updatePreviousSelection()
     }
