+++ lib/operator-insert.js
@@ -91,7 +91,7 @@ class ActivateInsertModeBase extends Operator {
           blockwiseSelection.skipNormalization()
         }
 
-        const disposable = this.vimState.modeManager.preemptWillDeactivateMode(({mode}) => {
+        const disposable = this.vimState.preemptWillDeactivateMode(({mode}) => {
           if (mode !== "insert") return
           disposable.dispose()
 
@@ -168,7 +168,7 @@ class ActivateReplaceMode extends ActivateInsertMode {
       }
     })
 
-    const disposable = this.vimState.modeManager.preemptWillDeactivateMode(({mode}) => {
+    const disposable = this.vimState.preemptWillDeactivateMode(({mode}) => {
       if (mode !== "insert") return
       disposable.dispose()
       onWillInsertDisposable.dispose()
