+++ lib/operator.js
@@ -47,7 +47,7 @@ class Operator extends Base {
   }
 
   // Two checkpoint for different purpose
-  // - one for undo(handled by modeManager)
+  // - one for undo
   // - one for preserve last inserted text
   createBufferCheckpoint(purpose) {
     this.bufferCheckpointByPurpose[purpose] = this.editor.createCheckpoint()
@@ -127,7 +127,7 @@ class Operator extends Base {
     if (this.selectPersistentSelectionIfNecessary()) {
       // [FIXME] selection-wise is not synched if it already visual-mode
       if (this.mode !== "visual") {
-        this.vimState.modeManager.activate("visual", this.swrap.detectWise(this.editor))
+        this.vimState.activate("visual", this.swrap.detectWise(this.editor))
       }
     }
 
@@ -485,7 +485,7 @@ class TogglePresetOccurrence extends Operator {
     if (marker) {
       this.occurrenceManager.destroyMarkers([marker])
     } else {
-      const isNarrowed = this.vimState.modeManager.isNarrowed()
+      const isNarrowed = this.vimState.isNarrowed()
 
       let regex
       if (this.mode === "visual" && !isNarrowed) {
