+++ lib/vim-state.js
@@ -1,8 +1,7 @@
 let jQuery, focusInput
 
-const {Emitter, Disposable, CompositeDisposable, Range} = require("atom")
+const {Emitter, Disposable, CompositeDisposable} = require("atom")
 const settings = require("./settings")
-const ModeManager = require("./mode-manager")
 
 const LazyLoadedLibs = {}
 function lazyRequire(file) {
@@ -27,14 +26,6 @@ module.exports = class VimState {
   static forEach(fn) { return __vimStatesByEditor.forEach(fn) } // prettier-ignore
   static clear() { return __vimStatesByEditor.clear() } // prettier-ignore
 
-  // To modeManager
-  get mode() { return this.modeManager.mode } // prettier-ignore
-  get submode() { return this.modeManager.submode } // prettier-ignore
-  // FIXME: REMOVE THIS; DONT directly update submode just for skip normalization
-  set submode(submode) { this.modeManager.submode = submode } // prettier-ignore
-
-  isMode(...args) { return this.modeManager.isMode(...args) } // prettier-ignore
-  activate(...args) { this.modeManager.activate(...args) } // prettier-ignore
   flash(...args) { this.flashManager.flash(...args) } // prettier-ignore
   clearFlash() { this.__flashManager && this.flashManager.clearAllMarkers() } // prettier-ignore
   updateStatusBar() { this.statusBarManager.update(this.mode, this.submode) } // prettier-ignore
@@ -75,13 +66,15 @@ module.exports = class VimState {
     this.editorElement = editor.element
     this.statusBarManager = statusBarManager
     this.emitter = new Emitter()
-    this.subscriptions = new CompositeDisposable()
-    this.modeManager = new ModeManager(this)
+
+    this.mode = "insert" // Bare atom is not modal editor, thus it's `insert` mode.
+    this.submode = null
+
     this.previousSelection = {}
-    this.scrollAnimationEffect = null
+    this.scrollRequest = null
     this.ignoreSelectionChange = false
 
-    this.subscriptions.add(
+    this.subscriptions = new CompositeDisposable(
       this.observeMouse(),
       this.editor.onDidAddSelection(selection => this.reconcileVisualModeWithActualSelection()),
       this.editor.onDidChangeSelectionRange(event => this.reconcileVisualModeWithActualSelection())
@@ -142,7 +135,7 @@ module.exports = class VimState {
     })
   }
 
-  // All subscriptions here is celared on each operation finished.
+  // All subscriptions here is cleared on each operation finished.
   // -------------------------
   onDidChangeSearch(fn) { return this.subscribe(this.searchInput.onDidChange(fn)) } // prettier-ignore
   onDidConfirmSearch(fn) { return this.subscribe(this.searchInput.onDidConfirm(fn)) } // prettier-ignore
@@ -176,24 +169,33 @@ module.exports = class VimState {
   onDidConfirmSelectList(fn) { return this.subscribe(this.emitter.on("did-confirm-select-list", fn)) } // prettier-ignore
   onDidCancelSelectList(fn) { return this.subscribe(this.emitter.on("did-cancel-select-list", fn)) } // prettier-ignore
 
-  onWillActivateMode(fn) { return this.subscribe(this.modeManager.onWillActivateMode(fn)) } // prettier-ignore
-  onDidActivateMode(fn) { return this.subscribe(this.modeManager.onDidActivateMode(fn)) } // prettier-ignore
-  onWillDeactivateMode(fn) { return this.subscribe(this.modeManager.onWillDeactivateMode(fn)) } // prettier-ignore
-  preemptWillDeactivateMode(fn) { return this.subscribe(this.modeManager.preemptWillDeactivateMode(fn)) } // prettier-ignore
-  onDidDeactivateMode(fn) { return this.subscribe(this.modeManager.onDidDeactivateMode(fn)) } // prettier-ignore
-
   // Events
   // -------------------------
-  onDidFailToPushToOperationStack(fn) {
-    return this.emitter.on("did-fail-to-push-to-operation-stack", fn)
-  }
-  emitDidFailToPushToOperationStack() {
-    this.emitter.emit("did-fail-to-push-to-operation-stack")
+  onWillActivateMode(fn) { return this.emitter.on("will-activate-mode", fn) } // prettier-ignore
+  onDidActivateMode(fn) { return this.emitter.on("did-activate-mode", fn) } // prettier-ignore
+  onWillDeactivateMode(fn) { return this.emitter.on("will-deactivate-mode", fn) } // prettier-ignore
+  preemptWillDeactivateMode(fn) { return this.emitter.preempt("will-deactivate-mode", fn) } // prettier-ignore
+  onDidDeactivateMode(fn) { return this.emitter.on("did-deactivate-mode", fn) } // prettier-ignore
+
+  get modeManager() {
+    const self = this
+    if (!this.modeManagerStub) {
+      this.modeManagerStub = {
+        get onWillActivateMode() { return self.onWillActivateMode.bind(self) }, // prettier-ignore
+        get onDidActivateMode() { return self.onDidActivateMode.bind(self) }, // prettier-ignore
+        get onWillDeactivateMode() { return self.onWillDeactivateMode.bind(self) }, // prettier-ignore
+        get preemptWillDeactivateMode() { return self.preemptWillDeactivateMode.bind(self) }, // prettier-ignore
+        get onDidDeactivateMode() { return self.onDidDeactivateMode.bind(self) }, // prettier-ignore
+      }
+    }
+    return this.modeManagerStub
   }
 
-  onDidDestroy(fn) {
-    return this.emitter.on("did-destroy", fn)
-  }
+  onDidFailToPushToOperationStack(fn) { return this.emitter.on("did-fail-to-push-to-operation-stack", fn) } // prettier-ignore
+  emitDidFailToPushToOperationStack() { this.emitter.emit("did-fail-to-push-to-operation-stack") } // prettier-ignore
+  onDidDestroy(fn) { return this.emitter.on("did-destroy", fn) } // prettier-ignore
+  onDidSetInputChar(fn) { return this.emitter.on("did-set-input-char", fn) } // prettier-ignore
+  emitDidSetInputChar(char) { this.emitter.emit("did-set-input-char", char) } // prettier-ignore
 
   // * `fn` {Function} to be called when mark was set.
   //   * `name` Name of mark such as 'a'.
@@ -205,13 +207,6 @@ module.exports = class VimState {
   //   onDidSetMark ({name, bufferPosition}) -> do something..
   onDidSetMark(fn) {
     return this.emitter.on("did-set-mark", fn)
-  }
-
-  onDidSetInputChar(fn) {
-    return this.emitter.on("did-set-input-char", fn)
-  }
-  emitDidSetInputChar(char) {
-    this.emitter.emit("did-set-input-char", char)
   }
 
   isAlive() {
@@ -461,6 +456,136 @@ module.exports = class VimState {
         if (onFinish) onFinish()
       },
     })
+  }
+
+  // Mode Managerment
+  //--------------------------
+  isMode(mode, submode) {
+    return mode === this.mode && (submode ? submode === this.submode : true)
+  }
+
+  // Use this method to change mode, DONT use other direct method.
+  // ---------------------------------
+  activate(newMode, newSubmode = null) {
+    // Avoid odd state(=visual-mode but selection is empty)
+    if (newMode === "visual" && this.editor.isEmpty()) return
+    this.ignoreSelectionChange = true
+
+    this.emitter.emit("will-activate-mode", {mode: newMode, submode: newSubmode})
+
+    if (newMode === "visual" && newSubmode && newSubmode === this.submode) {
+      newMode = "normal"
+      newSubmode = null
+    }
+
+    if (newMode !== this.mode) {
+      this.emitter.emit("will-deactivate-mode", {mode: this.mode, submode: this.submode})
+      if (this.modeDeactivator) {
+        this.modeDeactivator.dispose()
+        this.modeDeactivator = null
+      }
+      this.emitter.emit("did-deactivate-mode", {mode: this.mode, submode: this.submode})
+    }
+
+    if (newMode === "normal") this.activateNormalMode()
+    else if (newMode === "insert") this.editorElement.component.setInputEnabled(true)
+    else if (newMode === "visual") this.modeDeactivator = this.activateVisualMode(newSubmode)
+
+    this.editorElement.classList.remove(`${this.mode}-mode`)
+    this.editorElement.classList.remove(this.submode)
+
+    const oldMode = this.mode
+    this.mode = newMode
+    this.submode = newSubmode
+
+    if (oldMode === "visual" || this.mode === "visual") this.updateNarrowedState()
+
+    // Prevent swrap from loaded on initial mode-setup on startup.
+    if (this.mode === "visual") {
+      this.updatePreviousSelection()
+    } else {
+      if (this.__swrap) this.swrap.clearProperties(this.editor)
+    }
+
+    this.editorElement.classList.add(`${this.mode}-mode`)
+    if (this.submode) this.editorElement.classList.add(this.submode)
+
+    this.statusBarManager.update(this.mode, this.submode)
+    if (this.mode === "visual" || this.__cursorStyleManager) {
+      this.cursorStyleManager.refresh()
+    }
+
+    this.emitter.emit("did-activate-mode", {mode: this.mode, submode: this.submode})
+    this.ignoreSelectionChange = false
+  }
+
+  // Normal
+  // -------------------------
+  activateNormalMode() {
+    this.reset()
+    // Component is not necessary avaiable see #98.
+    if (this.editorElement.component) {
+      this.editorElement.component.setInputEnabled(false)
+    }
+
+    // In visual-mode, cursor can place at EOL. move left if cursor is at EOL
+    // We should not do this in visual-mode deactivation phase.
+    // e.g. `A` directly shift from visua-mode to `insert-mode`, and cursor should remain at EOL.
+    for (const cursor of this.editor.getCursors()) {
+      // Don't use utils moveCursorLeft to skip require('./utils') for faster startup.
+      if (cursor.isAtEndOfLine() && !cursor.isAtBeginningOfLine()) {
+        const {goalColumn} = cursor
+        cursor.moveLeft()
+        if (goalColumn != null) cursor.goalColumn = goalColumn
+      }
+    }
+  }
+
+  // Visual
+  // -------------------------
+  // We treat all selection is initially NOT normalized
+  //
+  // 1. First we normalize selection
+  // 2. Then update selection orientation(=wise).
+  //
+  // Regardless of selection is modified by vmp-command or outer-vmp-command like `cmd-l`.
+  // When normalize, we move cursor to left(selectLeft equivalent).
+  // Since Vim's visual-mode is always selectRighted.
+  //
+  // - un-normalized selection: This is the range we see in visual-mode.( So normal visual-mode range in user perspective ).
+  // - normalized selection: One column left selcted at selection end position
+  // - When selectRight at end position of normalized-selection, it become un-normalized selection
+  //   which is the range in visual-mode.
+  activateVisualMode(submode) {
+    const swrap = this.swrap
+    swrap.saveProperties(this.editor)
+    swrap.normalize(this.editor)
+
+    for (const $selection of swrap.getSelections(this.editor)) {
+      $selection.applyWise(submode)
+    }
+    if (submode === "blockwise") this.getLastBlockwiseSelection().autoscroll()
+
+    return new Disposable(() => {
+      swrap.normalize(this.editor)
+      if (this.submode === "blockwise") swrap.setReversedState(this.editor, true)
+      for (const selection of this.editor.getSelections()) {
+        selection.clear({autoscroll: false})
+      }
+    })
+  }
+
+  // Narrowed selection
+  // -------------------------
+  updateNarrowedState() {
+    const isSingleRowSelection = this.isMode("visual", "blockwise")
+      ? this.getLastBlockwiseSelection().isSingleRow()
+      : this.swrap(this.editor.getLastSelection()).isSingleRow()
+    this.editorElement.classList.toggle("is-narrowed", !isSingleRowSelection)
+  }
+
+  isNarrowed() {
+    return this.editorElement.classList.contains("is-narrowed")
   }
 
   // Other
